Clean up ALL the things
Make a lot of things more unified.
There are three ways to get IPs to check. You specify an interface, a
range, or a file to read from. All of these output the same thing, so it
doesn't affect later steps.

You also specify something to check. You can specify ping or portscan,
both of which can output a list of ips.

You can also specify an output file and a format. Format is either text
(newline separated ip's) or json. The format affects both how files are
read in (If that's how you're getting ip's), as well as how it outputs
the files.

If you specify an outfile, it will write either a newline separated list
of ip's (if you specify text), or a json array. If an outfile isn't
specified, a newline separated list will just be echo'd to STDOUT.